split test-invoker
